+++ guizero/Combo.py
@@ -1,10 +1,23 @@
 from tkinter import OptionMenu, StringVar, END, _setit
+from .mixins import WidgetMixin
 from .tkmixins import ScheduleMixin, DestroyMixin, EnableMixin, FocusMixin, DisplayMixin, ReprMixin
 from . import utilities as utils
 
-class Combo(ScheduleMixin, DestroyMixin, EnableMixin, FocusMixin, DisplayMixin, ReprMixin):
+class Combo(
+    WidgetMixin,
+    ScheduleMixin, 
+    DestroyMixin, 
+    EnableMixin, 
+    FocusMixin, 
+    DisplayMixin, 
+    ReprMixin):
 
     def __init__(self, master, options, selected=None, command=None, grid=None, align=None):
+
+        self._master = master
+        self._grid = grid
+        self._align = align
+        self._visible = True
 
         # Maintain a list of options (as strings, to avoid problems comparing)
         self._options = [str(x) for x in options]