+++ guizero/RadioButton.py
@@ -1,13 +1,26 @@
 from tkinter import Radiobutton
+from .mixins import WidgetMixin
 from .tkmixins import ScheduleMixin, DestroyMixin, EnableMixin, FocusMixin, DisplayMixin, ReprMixin
 from . import utilities as utils
 
 # Only to be used as part of button group
 # unless you want to externally create a controlling variable StringVar()
 # Does NOT create an internal StringVar() as this is handled by the ButtonGroup class
-class RadioButton(ScheduleMixin, DestroyMixin, EnableMixin, FocusMixin, DisplayMixin, ReprMixin):
+class RadioButton(
+    WidgetMixin, 
+    ScheduleMixin, 
+    DestroyMixin, 
+    EnableMixin, 
+    FocusMixin, 
+    DisplayMixin, 
+    ReprMixin):
 
     def __init__(self, master, text, value, variable, command=None, grid=None, align=None):
+
+        self._master = master
+        self._grid = grid
+        self._align = align
+        self._visible = True
 
         self.description = "[RadioButton] object with option=\"" + str(text) + "\" value=\"" + str(value) + "\""
         self._text = text